+++ src/rays.cpp
@@ -25,11 +25,11 @@
 
 class VisionRay : public Ray {
 public:
-  double *pixel;
-  double *zBufferPixel;
-  double startRefractIndex;
+  float *pixel;
+  float *zBufferPixel;
+  float startRefractIndex;
 
-  VisionRay(double *pixel_in, double *zBufferPixel_in, double *pos) {
+  VisionRay(float *pixel_in, float *zBufferPixel_in, float *pos) {
     type = 0;
     sumTs = 0;
     pixel = pixel_in;
@@ -37,7 +37,7 @@ public:
     weight[0] = weight[1] = weight[2] = 1.0;
     blasCopy(pos, position);
     startRefractIndex = 1.0;
-    intersectionT = std::numeric_limits<double>::infinity();
+    intersectionT = std::numeric_limits<float>::infinity();
   }
 
   VisionRay(VisionRay *r, int rayType) {
@@ -46,24 +46,24 @@ public:
     pixel = r->pixel;
     zBufferPixel = NULL;
     startRefractIndex = r->intersectionMaterial->interior;
-    intersectionT = std::numeric_limits<double>::infinity();
+    intersectionT = std::numeric_limits<float>::infinity();
     blasCopy(r->intersectionPoint, position);
 
-    double n[3];
+    float n[3];
     blasCopy(r->direction, direction);
     blasCopy(r->intersectionNormal, n);
-    double cosT1 = blasDot(n, direction);
+    float cosT1 = blasDot(n, direction);
 
     // Refraction Rays
     if(rayType == 1) {
-      double alpha = 1 - r->intersectionMaterial->opacy;
-      double filter[3];
+      float alpha = 1 - r->intersectionMaterial->opacy;
+      float filter[3];
       filter[0] = filter[1] = filter[2] =  r->intersectionMaterial->filter;
       filter[0] *= 1 - r->intersectionMaterial->color[0];
       filter[1] *= 1 - r->intersectionMaterial->color[1];
       filter[2] *= 1 - r->intersectionMaterial->color[2];
 
-      double nRatio = r->intersectionMaterial->interior;
+      float nRatio = r->intersectionMaterial->interior;
 
       if(nRatio != 1.0) {
         if(cosT1 < 0)
@@ -71,14 +71,14 @@ public:
         else
           blasInvert(n, n);
 
-        double sinT2 = nRatio*nRatio * (1 - cosT1*cosT1);
+        float sinT2 = nRatio*nRatio * (1 - cosT1*cosT1);
 
         if(sinT2 > 1) {
           // Total Reflection
           weight[0] = weight[1] = weight[2] = 0.0;
         }
         else {
-          double cosT2 = sqrt(1 - sinT2);
+          float cosT2 = sqrt(1 - sinT2);
           blasScale(direction, nRatio, direction);
           if(cosT1 > 0)
             blasScale(n, nRatio*cosT1 - cosT2, n);
@@ -110,7 +110,7 @@ public:
       blasSubstract(direction, n, direction);
       blasNormalize(direction);
 
-      double reflectiveAlpha = r->intersectionMaterial->reflection;
+      float reflectiveAlpha = r->intersectionMaterial->reflection;
       weight[0] = r->weight[0] * reflectiveAlpha;
       weight[1] = r->weight[1] * reflectiveAlpha;
       weight[2] = r->weight[2] * reflectiveAlpha;
@@ -124,22 +124,22 @@ public:
 
 class ShadowRay : public Ray {
 public:
-  ShadowRay(double *pos, double *dir) {
+  ShadowRay(float *pos, float *dir) {
     type = -1;
     sumTs = 0;
     blasCopy(pos, position);
     blasCopy(dir, direction);
     weight[0] = weight[1] = weight[2] = 1.0;
-    intersectionT = std::numeric_limits<double>::infinity();
+    intersectionT = std::numeric_limits<float>::infinity();
   }
 
   void nextStep() {
     sumTs += intersectionT;
     blasCopy(intersectionPoint, position);
-    intersectionT = std::numeric_limits<double>::infinity();
+    intersectionT = std::numeric_limits<float>::infinity();
     // Shadow Opacy + Color Filtering
-    double alpha = 1 - intersectionMaterial->opacy;
-    double filter[3];
+    float alpha = 1 - intersectionMaterial->opacy;
+    float filter[3];
     filter[0] = filter[1] = filter[2] =  intersectionMaterial->filter;
     filter[0] *= 1 - intersectionMaterial->color[0];
     filter[1] *= 1 - intersectionMaterial->color[1];
